Handle empty ascii responses from NXAPI
If the command passed to an NX API request does not produce any return
output then the API returns an empty hash. This appears to be unexpected
for request types of cli_show_ascii, so if an empty value is received
and the type is cli_show_ascii then we should return an empty string.

The previous behavior of returning an empty hash caused unhandled
failures in the node_utils gem as well.